abstract out DefaultLightblueSingleValueResponse
